Fix VPN pump return type sign-extension bug
Change alya_vpn_pump_client_vpn/server_vpn to return int64_t instead of int, so error codes (-1) aren't sign-extended incorrectly across the FFI boundary as unsigned 32-bit values. Update ffi.alya bindings to i64, adjust callers in local_proxy.alya and forwarder.alya to check for -1/4294967295 as disconnect signals, and improve client disconnect logging with address/port and max client count. Add e2e test coverage for active client tracking and pump error signaling.

+++ c/proc_resolver.h
@@ -175,13 +175,13 @@ int alya_vpn_open_client_channel(int vpn_sock, int channel_id, const char *host,
 // Reads AV02 binary frames from vpn_sock and forwards raw payload directly to app sockets;
 // Reads raw application bytes from active channels, encrypts into AV02 frames, and forwards to vpn_sock.
 // Returns 1 if activity, 0 if idle, -1 if vpn_sock closed/error.
-int alya_vpn_pump_client_vpn(int vpn_sock);
+int64_t alya_vpn_pump_client_vpn(int vpn_sock);
 
 // High-speed native server tunnel pump:
 // Reads AV02 binary frames from client_sock, handles CONNECT_REQ / DATA / CLOSE / PING;
 // Reads raw destination bytes from active server channels, encrypts into AV02 frames, forwards to client_sock.
 // Returns 1 if activity, 0 if idle, -1 if client_sock closed/error.
-int alya_vpn_pump_server_vpn(int client_sock);
+int64_t alya_vpn_pump_server_vpn(int client_sock);
 
 // Automatically configures or restores the OS system proxy (127.0.0.1:port)
 void alya_vpn_set_system_proxy(int enable, int port);
